+++ calc-worker-dotnet/.gitignore
@@ -1,3 +1,3 @@
-*/bin/
-*/obj/
+bin/
+obj/
 .vscode/
